Add files via upload
#Analysis of the most significant factors affecting education results

Education is one of the fundamental pillars of the development of a country. Having well educated children is a guarantee for technological and economic progress in the next decades. However, we know there are many factors that may affect students performance. Detecting the causes behind academic failure is the best way to tackle them form their origin to have an equal opportunity society.

I propose to take advantage of the numerous data and statistics available in the US Department of Education webpage to gain some insight of variables influencing education results. These data include State Assessments results, teachers/administrators characteristics, demographics, student outcomes and graduation rates for a great number of schools and high schools and many other aspects which can be of great help in this kind of study.

In the supporting plots I produced for the challenge, I analyzed data with information abo